Rename to viirs_edr_fires
