+++ satpy/readers/viirs_edr_fires.py
